spacesNg/requirements.txt
asgiref==3.3.4
boto3==1.17.101
botocore==1.20.101
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-storages==1.11.1
django-taggit==1.4.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.1
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
